Add support for concrete types
Now we can symbolically write literal values. To combine effort,
CDataType will have a getType() method which can be used for
constructors.

An iter() method is implemented which recursively traverses the
BinarySymExpr in a tree-like fashion. Further, we should move declare()
from Draw to a separate trait and Concrete types should not implement
declare().

ChainedBSEs are vector of many BSEs joined together by a logical binary
operator. The join() method is unfinished. Although, it should utilize
the iter() method for traversing in a tree-like manner.